move repack functions out of resourceprocessing
