+++ src/rsrc_mgr/resourceprocessing.c
@@ -71,15 +71,6 @@ GNU licenses can be found at http://www.gnu.org/licenses/.
 #define ENOATTR ENODATA
 #endif
 
-typedef struct repackstreamer {
-   // synchronization and access control
-   pthread_mutex_t lock;
-   // state info
-   size_t streamcount;
-   DATASTREAM* streamlist;
-   char* streamstatus;
-}* REPACKSTREAMER;
-
 typedef struct streamwalker {
    // initialization info
    marfs_position pos;
@@ -111,231 +102,6 @@ typedef struct streamwalker {
    // rebuild info
    opinfo*     rbldops;      // rebuild operation list
 }* streamwalker;
-
-
-//   -------------   REPACKSTREAMER FUNCTIONS    -------------
-
-/**
- * Initialize a new repackstreamer
- * @return REPACKSTREAMER : New repackstreamer, or NULL on failure
- */
-REPACKSTREAMER repackstreamer_init(void) {
-   // allocate a new struct
-   REPACKSTREAMER repackst = malloc(sizeof(*repackst));
-   if (repackst == NULL) {
-      LOG(LOG_ERR, "Failed to allocate new repackstreamer\n");
-      return NULL;
-   }
-   // populate all struct elements
-   if (pthread_mutex_init(&repackst->lock, NULL)) {
-      LOG(LOG_ERR, "Failed to initialize lock\n");
-      free(repackst);
-      return NULL;
-   }
-   repackst->streamcount = 10;
-   repackst->streamlist = calloc(10, sizeof(DATASTREAM));
-   if (repackst->streamlist == NULL) {
-      LOG(LOG_ERR, "Failed to initialize streamlist\n");
-      pthread_mutex_destroy(&repackst->lock);
-      free(repackst);
-      return NULL;
-   }
-   repackst->streamstatus = calloc(10, sizeof(char));
-   if (repackst->streamstatus == NULL) {
-      LOG(LOG_ERR, "Failed to initialize streamstatus\n");
-      free(repackst->streamlist);
-      pthread_mutex_destroy(&repackst->lock);
-      free(repackst);
-      return NULL;
-   }
-   return repackst;
-}
-
-/**
- * Checkout a repack datastream
- * @param REPACKSTREAMER repackst : Repackstreamer to checkout from
- * @return DATASTREAM* : Checked out datastream, or NULL on failure
- */
-DATASTREAM* repackstreamer_getstream(REPACKSTREAMER repackst) {
-   // check for NULL arg
-   if (repackst == NULL) {
-      LOG(LOG_ERR, "Received a NULL repackstreamer ref\n");
-      errno = EINVAL;
-      return NULL;
-   }
-   // acquire struct lock
-   if (pthread_mutex_lock(&repackst->lock)) {
-      LOG(LOG_ERR, "Failed to acquire repackstreamer lock\n");
-      return NULL;
-   }
-   // check for available datastreams
-   size_t index = 0;
-   for (; index < repackst->streamcount; index++) {
-      if (repackst->streamstatus[index] == 0) {
-         repackst->streamstatus[index] = 1;
-         pthread_mutex_unlock(&repackst->lock);
-         LOG(LOG_INFO, "Handing out available stream at position %zu\n", index);
-         return repackst->streamlist + index;
-      }
-   }
-   // no avaialable datastreams, so we must expand our allocation (double the current count)
-   LOG(LOG_INFO, "Expanding allocation to %zu streams\n", repackst->streamcount * 2);
-   DATASTREAM* newstlist = realloc(repackst->streamlist, sizeof(DATASTREAM) * (repackst->streamcount * 2));
-   if (newstlist == NULL) {
-      LOG(LOG_ERR, "Failed to reallocate streamlist to a length of %zu entries\n", repackst->streamcount * 2);
-      pthread_mutex_unlock(&repackst->lock);
-      return NULL;
-   }
-   repackst->streamlist = newstlist; // might end up leaving this expanded, but that's fine
-   char* newststatus = realloc(repackst->streamstatus, sizeof(char) * (repackst->streamcount * 2));
-   if (newststatus == NULL) {
-      LOG(LOG_ERR, "Failed to reallocate streamstatus to a length of %zu entries\n", repackst->streamcount * 2);
-      pthread_mutex_unlock(&repackst->lock);
-      return NULL;
-   }
-   repackst->streamstatus = newststatus;
-   repackst->streamcount *= 2;
-   // zero out new allocation
-   size_t newpos = index; // cache the lowest, newly-allocated index
-   for (; index < repackst->streamcount; index++) {
-      repackst->streamlist[index] = NULL;
-      repackst->streamstatus[index] = 0;
-   }
-   // hand out a newly-allocated stream
-   repackst->streamstatus[newpos] = 1;
-   pthread_mutex_unlock(&repackst->lock);
-   LOG(LOG_INFO, "Handing out newly-allocated position %zu\n", newpos);
-   return repackst->streamlist + newpos;
-}
-
-/**
- * Return a previously checked out repack datastream
- * @param REPACKSTREAMER repackst : Repackstreamer to return to
- * @param DATASTREAM* stream : Repack datastream to return
- * @return int : Zero on success, or -1 on failure
- */
-int repackstreamer_returnstream(REPACKSTREAMER repackst, DATASTREAM* stream) {
-   // check for NULL args
-   if (repackst == NULL) {
-      LOG(LOG_ERR, "Received a NULL repackstreamer ref\n");
-      errno = EINVAL;
-      return -1;
-   }
-   if (stream == NULL) {
-      LOG(LOG_ERR, "Received a NULL stream ref\n");
-      errno = EINVAL;
-      return -1;
-   }
-   // acquire struct lock
-   if (pthread_mutex_lock(&repackst->lock)) {
-      LOG(LOG_ERR, "Failed to acquire repackstreamer lock\n");
-      return -1;
-   }
-   // calculate the corresponding index of this stream
-   size_t index = (size_t)(stream - repackst->streamlist);
-   // sanity check the result
-   if (index >= repackst->streamcount  ||  stream < repackst->streamlist) {
-      LOG(LOG_ERR, "Returned stream is not a member of allocated list\n");
-      pthread_mutex_unlock(&repackst->lock);
-      return -1;
-   }
-   // ensure the stream was indeed passed out
-   if (repackst->streamstatus[index] != 1) {
-      LOG(LOG_ERR, "Returned stream %zu was not currently active\n", index);
-      pthread_mutex_unlock(&repackst->lock);
-      return -1;
-   }
-   // update status and return
-   repackst->streamstatus[index] = 0;
-   pthread_mutex_unlock(&repackst->lock);
-   LOG(LOG_INFO, "Stream %zu has been returned\n", index);
-   return 0;
-}
-
-/**
- * Terminate the given repackstreamer and close all associated datastreams
- * @param REPACKSTREAMER repackst : Repackstreamer to close
- * @return int : Zero on success, or -1 on failure
- */
-int repackstreamer_complete(REPACKSTREAMER repackst) {
-   // check for NULL arg
-   if (repackst == NULL) {
-      LOG(LOG_ERR, "Received a NULL repackstreamer ref\n");
-      errno = EINVAL;
-      return -1;
-   }
-   // acquire struct lock
-   if (pthread_mutex_lock(&repackst->lock)) {
-      LOG(LOG_ERR, "Failed to acquire repackstreamer lock\n");
-      return -1;
-   }
-   // iterate over all streams
-   int retval = 0;
-   char prevactive = 1;
-   size_t index = 0;
-   for (; index < repackst->streamcount; index++) {
-      if (*(repackst->streamlist + index) != NULL) {
-         // minor sanity check, not even certain that this is a true failure
-         if (prevactive == 0) {
-            LOG(LOG_WARNING, "Encountered active stream at index %zu with previous stream being NULL\n");
-         }
-         // close the active stream
-         int closeres = datastream_close(repackst->streamlist + index);
-         if (closeres) {
-            LOG(LOG_ERR, "Failed to close repack stream %zu\n", index);
-            if (retval == 0) { retval = closeres; }
-         }
-      }
-      else { prevactive = 0; }
-   }
-   // free all allocations
-   free(repackst->streamstatus);
-   free(repackst->streamlist);
-   pthread_mutex_unlock(&repackst->lock);
-   pthread_mutex_destroy(&repackst->lock);
-   free(repackst);
-   return retval;
-}
-
-/**
- * Abort the given repackstreamer, bypassing all locks and releasing all datastreams
- * @param REPACKSTREAMER repackst : Repackstreamer to abort
- * @return int : Zero on success, or -1 on failure
- */
-int repackstreamer_abort(REPACKSTREAMER repackst) {
-   // check for NULL arg
-   if (repackst == NULL) {
-      LOG(LOG_ERR, "Received a NULL repackstreamer ref\n");
-      errno = EINVAL;
-      return -1;
-   }
-   // don't bother acquiring the lock
-   // iterate over all streams
-   int retval = 0;
-   char prevactive = 1;
-   size_t index = 0;
-   for (; index < repackst->streamcount; index++) {
-      if (*(repackst->streamlist + index) != NULL) {
-         // minor sanity check, not even certain that this is a true failure
-         if (prevactive == 0) {
-            LOG(LOG_WARNING, "Encountered active stream at index %zu with previous stream being NULL\n");
-         }
-         // release the active stream
-         int closeres = datastream_release(repackst->streamlist + index);
-         if (closeres) {
-            LOG(LOG_ERR, "Failed to release repack stream %zu\n", index);
-            if (retval == 0) { retval = closeres; }
-         }
-      }
-      else { prevactive = 0; }
-   }
-   // free all allocations
-   free(repackst->streamstatus);
-   free(repackst->streamlist);
-   pthread_mutex_destroy(&repackst->lock);
-   free(repackst);
-   return retval;
-}
 
 
 //   -------------   INTERNAL FUNCTIONS    -------------
